pyproject.toml
[tool.poetry]
name = "ta_autograder"
version = "0.1.0"
description = "Experiment to determine effectiveness of LLMs for grading"
authors = ["TanglyEagle7718"]

[tool.poetry.dependencies]
python = "^3.10"
google-generativeai = "^0.8.4"
google-genai = "^1.5.0"
pandas = "^2.2.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
